refactor: extract api.ts orchestration helpers into api/ (#585)
Pure refactor — no logic changes. The 957-line `api.ts` orchestration file is
slimmed to ~416 lines by extracting six focused setup helpers under `api/`:

- middleware.ts (125)  — audit log, auth, admin middleware factories
- auth-routes.ts (196) — /version, /status, /signup, /login, /auth/token
- oauth-routes.ts (110) — Garmin / Strava / Oura connect+disconnect endpoints
- sync-setup.ts (160)  — `/sync` router wiring + per-provider sync transforms
- webhooks-setup.ts (138) — Strava + Oura push integration setup
- rest-routes.ts (115) — per-domain REST router mounts

The remaining `api.ts` keeps client construction (oura/garmin/strava), job
queues (pg-boss, deduction, strava, geocode), and server lifecycle
(httpd.listen + graceful shutdown).
